Have /departures endpoints adjust UTC timestamps to DST
The /departures endpoints should always return a UTC timestamp that accurately portrays the departure, at any time of the year, without the client having to adjust the time on their side (e.g. to account for DST).
In other words, if a departure is scheduled at 08:00 in a UTC+1 timezone that is subject to DST (e.g. CET), it should be returned as 07:00:00.000Z during winter time, but as 06:00:00.000Z during DST (aka summer time).